==> 0_analise_nomes_municipios.ipynb <==
import pandas as pd
import numpy as np

covid = pd.read_csv('dados/SUS/COVIDBR_2023_17jan2023.csv', sep=';')
eleicao18 = pd.read_csv('dados/TSE/2018_Bolsonaro.csv', sep=';')
eleicao22 = pd.read_csv('dados/TSE/2022_Bolsonaro.csv', sep=';')

covid = covid[['regiao', 'estado', 'municipio', 'data']]
covid = covid[covid['data'] == '2023-01-17']
covid = covid[~((covid.estado.isna()) | (covid.municipio.isna()))]
covid.reset_index(inplace=True)
covid = covid.drop(columns = ['index', 'data'])
covid['municipio_estado'] = (covid.municipio + ' / ' + covid.estado)
covid

import unicodedata

for i in range(len(covid.municipio_estado)):
    covid.municipio_estado[i] = ''.join(ch for ch in unicodedata.normalize('NFKD', covid.municipio_estado[i].upper())
        if not unicodedata.combining(ch))

covid

eleicao18

eleicao22

eleicao18 = eleicao18[['sg_uf','nm_municipio']]
eleicao22 = eleicao22[['sg_uf','nm_municipio']]
eleicao18['municipio_estado'] = eleicao18.nm_municipio + ' / ' + eleicao18.sg_uf
eleicao22['municipio_estado'] = eleicao22.nm_municipio + ' / ' + eleicao22.sg_uf
eleicao18 = eleicao18[eleicao18.sg_uf != 'ZZ']
eleicao22 = eleicao22[eleicao22.sg_uf != 'ZZ']
eleicao18.columns = ['uf_18', 'mun_18', 'mun_est_18']
eleicao22.columns = ['uf_22', 'mun_22', 'mun_est_22']

for i in range(len(eleicao18.mun_est_18)):
    eleicao18.mun_est_18[i] = ''.join(ch for ch in unicodedata.normalize('NFKD', eleicao18.mun_est_18[i]).upper() 
        if not unicodedata.combining(ch))


for i in range(len(eleicao22.mun_est_22)):
    eleicao22.mun_est_22[i] = ''.join(ch for ch in unicodedata.normalize('NFKD', eleicao22.mun_est_22[i]).upper() 
        if not unicodedata.combining(ch))



covid = covid.sort_values(by='municipio_estado')
covid.reset_index(inplace=True)
covid = covid.drop(columns=['index'])

eleicao18 = eleicao18.sort_values(by='mun_est_18')
eleicao18.reset_index(inplace=True)
eleicao18 = eleicao18.drop(columns=['index'])

eleicao22 = eleicao22.sort_values(by='mun_est_22')
eleicao22.reset_index(inplace=True)
eleicao22 = eleicao22.drop(columns=['index'])

dados = pd.DataFrame(columns=['covid', 'eleicao18', 'eleicao22'])

dados['covid'] = covid.municipio_estado
dados['eleicao18'] = eleicao18.mun_est_18
dados['eleicao22'] = eleicao22.mun_est_22

dados

### Optei por levar todos os nomes para o dado mais atual - 'eleicao22'

lista_22 = dados.eleicao22.to_list()
lista_18 = dados.eleicao18.to_list()
lista_cov = dados.covid.to_list()

dados['corr_covid'] = (~dados['covid'].isin(lista_22)).astype(int)
dados['22_covid'] = (~dados['eleicao22'].isin(lista_cov)).astype(int)

dados['corr_18'] = (~dados['eleicao18'].isin(lista_22)).astype(int)
dados['22_18'] = (~dados['eleicao22'].isin(lista_18)).astype(int)

dados['covid_18'] = (~dados['eleicao18'].isin(lista_cov)).astype(int)
dados['18_covid'] = (~dados['covid'].isin(lista_18)).astype(int)



dados['corrigir'] = dados['corr_covid'] + dados['22_covid'] + dados['corr_18'] + dados['22_18'] + dados['covid_18'] + dados['18_covid']

dados[dados['corrigir'] != 0]

covid	

corretor = list(['ASSU',
  'ALVORADA DO OESTE',
  'AREZ',
  'BOA SAUDE',
  'CAMACA',
  'DONA EUSEBIA',
  'ELDORADO DOS CARAJAS',
  'ESPIGAO DO OESTE',
  'GRACCHO CARDOSO',
  'GRAO-PARA',
  "OLHOS D'AGUA",
  'SANTA ISABEL DO PARA',
  'SAO LUIS DO PARAITINGA',
  'SAO THOME DAS LETRAS',
  'SEM PEIXE'
])

trocas = list(['ACU',
    'ALVORADA D'OESTE',
    'ARES',
    'JANUARIO CICCO',
    'CAMACAN',
    'DONA EUZEBIA',
    'ELDORADO DO CARAJAS',
    'ESPIGAO D'OESTE',
    'GRACHO CARDOSO',
    'GRAO PARA',
    'OLHOS-D'AGUA',
    'SANTA IZABEL DO PARA',
    'SAO LUIZ DO PARAITINGA',
    'SAO TOME DAS LETRAS',
    'SEM-PEIXE
])


eleicao18	

corretor = list(['AREZ',
  'TABOCAO',
  'QUIJINGUE'
])

trocas = list(['ARES'
  'FORTALEZA DO TABOCAO'
  'QUINJINGUE'
])

eleicao22



### Verificação após as correções (Rodar nos arquivos salvos)

n_covid = pd.read_csv('arquivos/covid_municipios.csv')
n_eleicao18 = pd.read_csv('arquivos/eleicao18.csv')
n_eleicao22 = pd.read_csv('arquivos/eleicao22.csv')

n_covid['municipio_estado'] = n_covid['municipio'] + ' / ' + n_covid['estado']
n_eleicao18['municipio_estado'] = n_eleicao18['municipio'] + ' / ' + n_eleicao18['estado']
n_eleicao22['municipio_estado'] = n_eleicao22['municipio'] + ' / ' + n_eleicao22['estado']

n_eleicao18 = n_eleicao18[n_eleicao18['estado'] != 'ZZ']
n_eleicao22 = n_eleicao22[n_eleicao22['estado'] != 'ZZ']

n_covid = n_covid.sort_values(by='municipio_estado')
n_covid.reset_index(inplace=True)
n_covid = n_covid.drop(columns=['index'])

n_eleicao18 = n_eleicao18.sort_values(by='municipio_estado')
n_eleicao18.reset_index(inplace=True)
n_eleicao18 = n_eleicao18.drop(columns=['index'])

n_eleicao22 = n_eleicao22.sort_values(by='municipio_estado')
n_eleicao22.reset_index(inplace=True)
n_eleicao22 = n_eleicao22.drop(columns=['index'])

n_dados = pd.DataFrame(columns=['covid', 'eleicao18', 'eleicao22'])

n_dados.covid = n_covid.municipio_estado
n_dados.eleicao18 = n_eleicao18.municipio_estado
n_dados.eleicao22 = n_eleicao22.municipio_estado


n_dados

lista_22 = n_dados.eleicao22.to_list()
lista_18 = n_dados.eleicao18.to_list()
lista_cov = n_dados.covid.to_list()

n_dados['corr_covid'] = (~n_dados['covid'].isin(lista_22)).astype(int)
n_dados['22_covid'] = (~n_dados['eleicao22'].isin(lista_cov)).astype(int)

n_dados['corr_18'] = (~n_dados['eleicao18'].isin(lista_22)).astype(int)
n_dados['22_18'] = (~n_dados['eleicao22'].isin(lista_18)).astype(int)

n_dados['covid_18'] = (~n_dados['eleicao18'].isin(lista_cov)).astype(int)
n_dados['18_covid'] = (~n_dados['covid'].isin(lista_18)).astype(int)

n_dados['corrigir'] = n_dados['corr_covid'] + n_dados['22_covid'] + n_dados['corr_18'] + n_dados['22_18'] + n_dados['covid_18'] + n_dados['18_covid']

n_dados



n_dados[n_dados['corrigir'] != 0]
