==> 1_covid_tratamento_dados.ipynb <==
import pandas as pd
import numpy as np

covid = pd.read_csv('dados/SUS/COVIDBR_2023_17jan2023.csv', sep=';')

covid

covid[(covid.estado.isna()) | (covid.municipio.isna())]

covid.columns

covid = covid.drop(columns=['coduf', 'codmun', 'codRegiaoSaude', 'nomeRegiaoSaude', 'semanaEpi',
       'casosNovos', 'obitosNovos', 'Recuperadosnovos', 'emAcompanhamentoNovos', 'interior/metropolitana']
       )

covid_municipios = covid[~((covid.estado.isna()) | (covid.municipio.isna()))]
covid_municipios = covid_municipios[covid_municipios.data =='2023-01-17']
covid_municipios.columns = ['regiao', 'estado', 'municipio_0', 'data', 'populacao', 'casosAcumulado', 'obitosAcumulado']
covid_municipios.populacao = covid_municipios.populacao.astype(int)
covid_municipios.reset_index(inplace=True)
covid_municipios = covid_municipios.drop(columns = ['index'])

for i in range(len(covid_municipios.municipio_0)):
    covid_municipios.municipio_0[i] = covid_municipios.municipio_0[i].upper()

covid_municipios = covid_municipios.drop(columns=['data'])
covid_municipios.describe(include='all')

import unicodedata

for i in range(len(covid_municipios.municipio_0)):
    covid_municipios.municipio_0[i] = ''.join(ch for ch in unicodedata.normalize('NFKD', covid_municipios.municipio_0[i]) 
        if not unicodedata.combining(ch))

covid_municipios

# Corrigir 
Na planilha COVID  mapear
 
corretor = list(['ASSU',
  'ALVORADA DO OESTE',
  'AREZ',
  'BOA SAUDE',
  'CAMACA',
  'DONA EUSEBIA',
  'ELDORADO DOS CARAJAS',
  'ESPIGAO DO OESTE',
  'GRACCHO CARDOSO',
  'GRAO-PARA',
  "OLHOS D'AGUA",
  'SANTA ISABEL DO PARA',
  'SAO LUIS DO PARAITINGA',
  'SAO THOME DAS LETRAS',
  'SEM PEIXE'
])

trocas = list(['ACU',
    'ALVORADA D'OESTE',
    'ARES',
    'JANUARIO CICCO',
    'CAMACAN',
    'DONA EUZEBIA',
    'ELDORADO DO CARAJAS',
    'ESPIGAO D'OESTE',
    'GRACHO CARDOSO',
    'GRAO PARA',
    'OLHOS-D'AGUA',
    'SANTA IZABEL DO PARA',
    'SAO LUIZ DO PARAITINGA',
    'SAO TOME DAS LETRAS',
    'SEM-PEIXE
])

corretor = list(['ASSU',
  'ALVORADA DO OESTE',
  'AREZ',
  'BOA SAUDE',
  'CAMACA',
  'DONA EUSEBIA',
  'ELDORADO DOS CARAJAS',
  'ESPIGAO DO OESTE',
  'GRACCHO CARDOSO',
  'GRAO-PARA',
  "OLHOS D'AGUA",
  'SANTA ISABEL DO PARA',
  'SAO LUIS DO PARAITINGA',
  'SAO THOME DAS LETRAS',
  'SEM PEIXE'
])

trocas = list(['ACU',
    "ALVORADA D'OESTE",
    'ARES',
    'JANUARIO CICCO',
    'CAMACAN',
    'DONA EUZEBIA',
    'ELDORADO DO CARAJAS',
    "ESPIGAO D'OESTE",
    'GRACHO CARDOSO',
    'GRAO PARA',
    "OLHOS-D'AGUA",
    'SANTA IZABEL DO PARA',
    'SAO LUIZ DO PARAITINGA',
    'SAO TOME DAS LETRAS',
    'SEM-PEIXE'
])

covid_municipios['municipio'] = covid_municipios['municipio_0']

for i in range(len(covid_municipios['municipio_0'])):
    if covid_municipios['municipio_0'][i] in trocas:
        for k in range(len(trocas)):
            if covid_municipios['municipio_0'][i] == trocas[k]:
               covid_municipios['municipio'][i] = corretor[k]
    else:
        covid_municipios['municipio'][i] = covid_municipios['municipio_0'][i]

covid_municipios = covid_municipios.drop(columns=['municipio_0'])
covid_municipios = covid_municipios[['regiao','estado','municipio','populacao','casosAcumulado','obitosAcumulado']]

covid_municipios = covid_municipios.sort_values(by='municipio')

covid_municipios.to_csv('arquivos/covid_municipios.csv', index=False)

regionalize = covid_municipios.groupby('estado').head(1)
df_regiao = regionalize.estado, regionalize.regiao
df_regiao = pd.DataFrame(df_regiao)
df_regiao = df_regiao.transpose()
df_regiao = df_regiao.reset_index()
df_regiao = df_regiao.drop(columns=['index'])
df_regiao
