==> 2_eleicao18_tratamento_dados.ipynb <==
import pandas as pd
import numpy as np

dadosJair = pd.read_csv('DADOS/TSE/2018_Bolsonaro.csv', sep=';')
dadosHaddad = pd.read_csv('DADOS/TSE/2018_Haddad.csv', sep=';')

dadosJair

dadosHaddad

verificador = dadosHaddad.describe().qt_votos_nom_validos

dadosJair.columns

dados = dadosJair[['sg_uf', 'nm_municipio', 'qt_votos_nom_validos']]

dados.columns = ['estado', 'municipio_0', 'votos_Bolsonaro']

dados['votos_Haddad'] = 0
dados

for i in range(dadosHaddad.shape[0]):
    if dados.estado[i] == dadosHaddad.sg_uf[i]:
        dados.votos_Haddad[i] = (dadosHaddad.qt_votos_nom_validos[i])
    

dados

dados.info()

dados.describe().votos_Haddad == verificador

dados['regiao'] = ''

regiao = {'RO':	'Norte',
'AC': 'Norte',
'AM': 'Norte',
'RR': 'Norte',
'PA': 'Norte',
'AP': 'Norte',
'TO': 'Norte',
'MA': 'Nordeste',
'PI': 'Nordeste',
'CE': 'Nordeste',
'RN': 'Nordeste',
'PB': 'Nordeste',
'PE': 'Nordeste',
'AL': 'Nordeste',
'SE': 'Nordeste',
'BA': 'Nordeste',
'MG': 'Sudeste',
'ES': 'Sudeste',
'RJ': 'Sudeste',
'SP': 'Sudeste',
'PR': 'Sul',
'SC': 'Sul',
'RS': 'Sul',
'MS': 'Centro-Oeste',
'MT': 'Centro-Oeste',
'GO': 'Centro-Oeste',
'DF': 'Centro-Oeste',
'ZZ': 'Exterior'}

dados['regiao'] = dados.estado.map(regiao)

dados

dados.describe(include='all')

dados = dados[['regiao','estado','municipio_0','votos_Bolsonaro','votos_Haddad']].sort_values(by='municipio_0')
dados

import unicodedata

for i in range(len(dados.municipio_0)):
    dados.municipio_0[i] = ''.join(ch for ch in unicodedata.normalize('NFKD', dados.municipio_0[i]) 
        if not unicodedata.combining(ch))

dados

# Corrigir 
Na planilha eleicao18  mapear

corretor = list(['AREZ',
  'TABOCAO',
  'QUIJINGUE'
])

trocas = list(['ARES'
  'FORTALEZA DO TABOCAO'
  'QUINJINGUE'
])


corretor = list(['AREZ',
  'TABOCAO',
  'QUIJINGUE'
])

trocas = list(['ARES',
  'FORTALEZA DO TABOCAO',
  'QUINJINGUE'
])


dados['municipio'] = dados['municipio_0']

for i in range(len(dados['municipio_0'])):
    if dados['municipio_0'][i] in trocas:
        for k in range(len(trocas)):
            if dados['municipio_0'][i] == trocas[k]:
               dados['municipio'][i] = corretor[k]
    else:
        dados['municipio'][i] = dados['municipio_0'][i]

dados = dados.drop(columns=['municipio_0'])
dados = dados[['regiao','estado','municipio','votos_Bolsonaro', 'votos_Haddad']]

dados = dados.sort_values(by='municipio')

dados.to_csv('arquivos/eleicao18.csv', index=False)
