==> 3_eleicao22_tratamento_dados.ipynb <==
import pandas as pd
import numpy as np

dadosJair = pd.read_csv('DADOS/TSE/2022_Bolsonaro.csv', sep=';')
dadosLula = pd.read_csv('DADOS/TSE/2022_Lula.csv', sep=';')

dadosJair

dadosLula

verificador = dadosLula.describe().qt_votos_nom_validos

dadosJair.columns

dados = dadosJair[['sg_uf', 'nm_municipio', 'qt_votos_nom_validos']]

dados.columns = ['estado', 'municipio', 'votos_Bolsonaro']

dados['votos_Lula'] = 0
dados

for i in range(dadosLula.shape[0]):
    if dados.estado[i] == dadosLula.sg_uf[i]:
        dados.votos_Lula[i] = (dadosLula.qt_votos_nom_validos[i])
    

dados

dados.info()

dados.describe().votos_Lula == verificador

dados['regiao'] = ''

regiao = {'RO':	'Norte',
'AC': 'Norte',
'AM': 'Norte',
'RR': 'Norte',
'PA': 'Norte',
'AP': 'Norte',
'TO': 'Norte',
'MA': 'Nordeste',
'PI': 'Nordeste',
'CE': 'Nordeste',
'RN': 'Nordeste',
'PB': 'Nordeste',
'PE': 'Nordeste',
'AL': 'Nordeste',
'SE': 'Nordeste',
'BA': 'Nordeste',
'MG': 'Sudeste',
'ES': 'Sudeste',
'RJ': 'Sudeste',
'SP': 'Sudeste',
'PR': 'Sul',
'SC': 'Sul',
'RS': 'Sul',
'MS': 'Centro-Oeste',
'MT': 'Centro-Oeste',
'GO': 'Centro-Oeste',
'DF': 'Centro-Oeste',
'ZZ': 'Exterior'}

dados['regiao'] = dados.estado.map(regiao)

dados

dados.describe(include='all')

dados = dados[['regiao','estado','municipio','votos_Bolsonaro','votos_Lula']].sort_values(by='municipio')
dados

import unicodedata


for i in range(len(dados.municipio)):
    dados.municipio[i] = ''.join(ch for ch in unicodedata.normalize('NFKD', dados.municipio[i]) 
        if not unicodedata.combining(ch))

dados = dados.sort_values(by='municipio')

dados.to_csv('arquivos/eleicao22.csv', index=False)
