==> 5_arquivo_Looker.ipynb <==
import pandas as pd
import numpy as np

n_covid = pd.read_csv('arquivos/covid_municipios.csv')
n_eleicao18 = pd.read_csv('arquivos/eleicao18.csv')
n_eleicao22 = pd.read_csv('arquivos/eleicao22.csv')

n_covid['municipio_estado'] = n_covid['municipio'] + ' / ' + n_covid['estado']
n_eleicao18['municipio_estado'] = n_eleicao18['municipio'] + ' / ' + n_eleicao18['estado']
n_eleicao22['municipio_estado'] = n_eleicao22['municipio'] + ' / ' + n_eleicao22['estado']

n_eleicao18 = n_eleicao18[n_eleicao18['estado'] != 'ZZ']
n_eleicao22 = n_eleicao22[n_eleicao22['estado'] != 'ZZ']

n_covid = n_covid.sort_values(by='municipio_estado')
n_covid.reset_index(inplace=True, drop=True)

n_eleicao18 = n_eleicao18.sort_values(by='municipio_estado')
n_eleicao18.reset_index(inplace=True, drop=True)

n_eleicao22 = n_eleicao22.sort_values(by='municipio_estado')
n_eleicao22.reset_index(inplace=True, drop=True)

n_covid.head(2)

n_eleicao18.head(2)

n_eleicao22.head(2)

eleicoes = pd.merge(n_eleicao18, n_eleicao22[['municipio_estado','votos_Bolsonaro','votos_Lula']], on='municipio_estado')

eleicoes

combinados = pd.merge(n_covid, eleicoes[['municipio_estado','votos_Bolsonaro_x','votos_Haddad','votos_Bolsonaro_y','votos_Lula']], on='municipio_estado')

combinados

combinados.columns = ['regiao', 'estado', 'municipio', 'populacao', 'casosAcumulado',
       'obitosAcumulado', 'municipio_estado', 'votos_Bolsonaro_18',
       'votos_Haddad', 'votos_Bolsonaro_22', 'votos_Lula']

nome_dos_estados = {
'GO': 'Goiás',
'MG' : 'Minas Gerais',
'PA' : 'Pará',
'CE' : 'Ceará',
'BA' : 'Bahia',
'PR' : 'Paraná ',
'SC' : 'Santa Catarina',
'PE' : 'Pernambuco',
'TO' : 'Tocantins',
'MA' : 'Maranhão',
'RN' : 'Rio Grande do Norte',
'PI' : 'Piauí',
'RS' : 'Rio Grande do Sul',
'MT' : 'Mato Grosso',
'AC' : 'Acre',
'SP' : 'São Paulo',
'ES' : 'Espírito Santo',
'AL' : 'Alagoas',
'PB' : 'Paraíba',
'MS' : 'Mato Grosso do Sul',
'RO' : 'Rondônia',
'RR' : 'Roraima',
'AM' : 'Amazonas',
'AP' : 'Amapá',
'SE' : 'Sergipe',
'RJ' : 'Rio de Janeiro',
'DF' : 'Distrito Federal'}

combinados['estado_nome'] = combinados.estado.map(nome_dos_estados)


combinados['ISO_code'] = 'BR-' + combinados.estado

combinados

combinados.to_csv('arquivos/dados_looker.csv', index=False)
